生徒について,CRUDの動作確認完了
